feat(nvim): add keybindings inside core folder
